requirements.txt
pydantic
typing
openai
requests
transformers
torch
mistralai
python-dotenv
requests